move test function to new file
